Create troll-bridge.rb
Made with Page 1-4 of The Book Of Ruby by No Starch Press. This is a simple program where you are crossing a bridge and a troll stops you and asks for his toll (30% of what money you have in your pockets.) If you don't have any money, or less then zero, the troll gives you a lolly because he's not a bad guy.